services/ml-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
python-jose==3.3.0
httpx==0.26.0
azure-identity==1.15.0
azure-storage-blob==12.19.0
databricks-sdk==0.17.0
databricks-sql-connector==3.0.3
numpy==1.26.3
pandas==2.1.4
scikit-learn==1.3.2
transformers==4.36.2
torch==2.1.2
redis==5.0.1
prometheus-client==0.19.0
python-dotenv==1.0.0
langdetect==1.0.9
textblob==0.17.1
